refactor: move 3d model api client to util
