+++ backend/build.gradle
@@ -33,10 +33,13 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-webflux'
 
 	// ✅ Spring Security + JWT
-	implementation 'org.springframework.boot:spring-boot-starter-security'
-	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-	runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
-	runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
+        implementation 'org.springframework.boot:spring-boot-starter-security'
+        implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+        runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
+        runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
+
+        // Needed for MockMultipartFile
+        implementation 'org.springframework:spring-test'
 
 	// ✅ DB
 	runtimeOnly 'com.h2database:h2'
@@ -46,7 +49,9 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 
 	// PDFBox 라이브러리 추가
-	implementation 'org.apache.pdfbox:pdfbox:2.0.30'
+        implementation 'org.apache.pdfbox:pdfbox:2.0.30'
+
+        testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
 
 // ✅ 테스트 실행 비활성화
